Create separate file with filter controllers
